+++ src/App.tsx
@@ -7,6 +7,7 @@ import React from 'react';
 import { VaultProvider } from './contexts/VaultContext';
 import { TabsProvider } from './contexts/TabsContext';
 import { ActivityProvider } from './contexts/ActivityContext';
+import { AppSettingsProvider } from './contexts/AppSettingsContext';
 import { ModalProvider } from './components/Modal';
 import { Layout } from './components/Layout';
 
@@ -15,9 +16,11 @@ export default function App() {
     <VaultProvider>
       <TabsProvider>
         <ActivityProvider>
-          <ModalProvider>
-            <Layout />
-          </ModalProvider>
+          <AppSettingsProvider>
+            <ModalProvider>
+              <Layout />
+            </ModalProvider>
+          </AppSettingsProvider>
         </ActivityProvider>
       </TabsProvider>
     </VaultProvider>